+++ packages/@webex/plugin-meetings/test/unit/spec/meetings/index.js
@@ -30,6 +30,10 @@ import {
   LOCUSINFO,
   EVENT_TRIGGERS,
 } from '../../../../src/constants';
+import CaptchaError from '@webex/plugin-meetings/src/common/errors/captcha-error';
+import { forEach } from 'lodash';
+import PasswordError from '@webex/plugin-meetings/src/common/errors/password-error';
+import PermissionError from '@webex/plugin-meetings/src/common/errors/permission';
 
 describe('plugin-meetings', () => {
   const logger = {
@@ -1211,6 +1215,63 @@ describe('plugin-meetings', () => {
               }
             );
           });
+        });
+
+        describe('rejected MeetingInfo.#fetchMeetingInfo - does not log for known Error types', () => {
+          forEach(
+            [
+              {
+                error: new CaptchaError(),
+                debugLogMessage:
+                  'Meetings:index#createMeeting --> Debug CaptchaError: Captcha is required. fetching /meetingInfo for creation.',
+              },
+              {
+                error: new PasswordError(),
+                debugLogMessage:
+                  'Meetings:index#createMeeting --> Debug PasswordError: Password is required, please use verifyPassword() fetching /meetingInfo for creation.',
+              },
+              {
+                error: new PermissionError(),
+                debugLogMessage:
+                  'Meetings:index#createMeeting --> Debug PermissionError: Not allowed to execute the function, some properties on server, or local client state do not allow you to complete this action. fetching /meetingInfo for creation.',
+              },
+              {
+                error: new Error(),
+                infoLogMessage: true,
+                debugLogMessage:
+                  'Meetings:index#createMeeting --> Debug Error fetching /meetingInfo for creation.',
+              },
+            ],
+            ({error, debugLogMessage, infoLogMessage}) => {
+              it('creates the meeting from a rejected meeting info fetch', async () => {
+                webex.meetings.meetingInfo.fetchMeetingInfo = sinon
+                  .stub()
+                  .returns(Promise.reject(error));
+
+                LoggerProxy.logger.debug = sinon.stub();
+                LoggerProxy.logger.info = sinon.stub();
+
+                const meeting = await webex.meetings.createMeeting('test destination', 'test type');
+
+                assert.instanceOf(
+                  meeting,
+                  Meeting,
+                  'createMeeting should eventually resolve to a Meeting Object'
+                );
+
+                assert.calledWith(LoggerProxy.logger.debug, debugLogMessage);
+
+                if (infoLogMessage) {
+                  assert.calledWith(
+                    LoggerProxy.logger.info,
+                    'Meetings:index#createMeeting --> Info Unable to fetch meeting info for test destination.'
+                  );
+                } else {
+                  assert.notCalled(LoggerProxy.logger.info);
+                }
+              });
+            }
+          );
         });
       });
     });